refactored DecoderType into a separate file, fixed minor issues with the code
