+++ app/src/main/java/org/jellyfin/mobile/player/interaction/DecoderType.kt
@@ -1,0 +1,9 @@
+package org.jellyfin.mobile.player.interaction
+
+/**
+ * Represents the type of decoder
+ */
+enum class DecoderType {
+    HARDWARE,
+    SOFTWARE,
+}
